+++ loginUI/src/navigation/types.ts
@@ -2,4 +2,6 @@ export type RootStackParamList = {
   Login: undefined;
   AuthOptions: undefined;
   MainService: undefined;
+  GetStartedSignup: undefined;
+  VerifyOtp: { phoneNumber: string; countryCode: string };
 };